Work on ensuring H1 overlay animation import -> export parity.
The goal is to be able to import a .jmo, be able to easily modify it in Blender, and then export it and have it work in-game.

An imported -> directly exported .jmo should ideally have the exact same resulting animation in-game, though the output file will most likely never match the input due to differences in floating-point precision and such.

This current code *works* for a given definition of the word, though it's expansive, wordy, and brittle, with bad perf. The singular animation I've tested so far is the Elite 'stand pistol aim-still' overlay animation, which seemingly exports perfectly and appears identical in-game.

More work to be done.